Adding support for the "automixmonitor" dial and queue options.
This works in much the same way as the automonitor, except that instead of using the monitor
app, it uses the mixmonitor app. By providing an 'x' or 'X' as a dial or queue option, a DTMF
sequence may be entered (as defined in features.conf) to start the one-touch mixmonitor.

This patch also introduces some new API calls to the audiohooks code for searching for an audiohook
by type and for searching for a running audiohook by type.

Big thanks to joetester for writing the initial patch, testing it and patiently waiting for it to 
be committed.

(closes issue #10185, reported and patched by xmarksthespot)



git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@90388 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ include/asterisk/app.h
@@ -344,7 +344,7 @@ unsigned int ast_app_separate_args(char *buf, char delim, char **array, int arra
  */
 struct ast_app_option {
 	/*! \brief The flag bit that represents this option. */
-	unsigned int flag;
+	uint64_t flag;
 	/*! \brief The index of the entry in the arguments array
 	  that should be used for this option's argument. */
 	unsigned int arg_index;
